// src/generated-ts-client.ts
export interface ActionRequest {
    farmId?: string;
    turbineId?: string;
    action?: string;
    value?: number | undefined;
    reason?: string | undefined;
    angle?: number | undefined;
}